requirements.txt
runpod
openai

# Whisper (transcription)
openai-whisper
torch
numpy

# Vidéo / audio
ffmpeg-python
moviepy
pydub

# Sous-titres avancés
pysubs2

# Pour télécharger la vidéo depuis une URL
requests
